backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic>=2.10.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Added for Day 2 features
email-validator>=2.2.0
# Day 3: Cloudflare R2 integration
boto3==1.34.0
